Extract ScalarImplementation class from ReflectionParametricScalar
This change is focused on extracting elements describing single
implementation of Scalar function from ReflectionParametricScalar.

The class ScalarImplementation was exctracted from ReflectionParametricScalar.Implementation.
This change reasoning is that ScalarImplementation keeps description
that may be successfully used outside of ReflectionParametricScalar and
the connection between those two classes is that ReflectionParametricScalar uses
ScalarImplementation.

For keeping changes small the class is extracted together with its own
parser (also part of ReflectionParametricScalar before) as a private subclass.

This is why to keep classes as short as simple as possible it was extracted.
