+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/ReflectionParametricScalar.java
@@ -15,88 +15,47 @@ package com.facebook.presto.operator.scalar;
 
 import com.facebook.presto.metadata.BoundVariables;
 import com.facebook.presto.metadata.FunctionRegistry;
-import com.facebook.presto.metadata.LongVariableConstraint;
-import com.facebook.presto.metadata.OperatorType;
 import com.facebook.presto.metadata.Signature;
 import com.facebook.presto.metadata.SignatureBinder;
 import com.facebook.presto.metadata.SqlScalarFunction;
-import com.facebook.presto.metadata.TypeVariableConstraint;
+import com.facebook.presto.operator.scalar.annotations.ScalarImplementation;
 import com.facebook.presto.operator.scalar.annotations.ScalarImplementationHeader;
-import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.PrestoException;
-import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.spi.type.TypeManager;
 import com.facebook.presto.spi.type.TypeSignature;
-import com.facebook.presto.type.Constraint;
-import com.facebook.presto.type.LiteralParameters;
 import com.facebook.presto.type.SqlType;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Iterables;
-import com.google.common.primitives.Primitives;
-
-import javax.annotation.Nullable;
 
 import java.lang.annotation.Annotation;
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodType;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import static com.facebook.presto.metadata.FunctionKind.SCALAR;
-import static com.facebook.presto.metadata.OperatorType.BETWEEN;
-import static com.facebook.presto.metadata.OperatorType.CAST;
-import static com.facebook.presto.metadata.OperatorType.EQUAL;
-import static com.facebook.presto.metadata.OperatorType.GREATER_THAN;
-import static com.facebook.presto.metadata.OperatorType.GREATER_THAN_OR_EQUAL;
-import static com.facebook.presto.metadata.OperatorType.HASH_CODE;
-import static com.facebook.presto.metadata.OperatorType.LESS_THAN;
-import static com.facebook.presto.metadata.OperatorType.LESS_THAN_OR_EQUAL;
-import static com.facebook.presto.metadata.OperatorType.NOT_EQUAL;
-import static com.facebook.presto.metadata.Signature.comparableTypeParameter;
-import static com.facebook.presto.metadata.Signature.internalOperator;
-import static com.facebook.presto.metadata.Signature.orderableTypeParameter;
-import static com.facebook.presto.metadata.Signature.typeVariable;
 import static com.facebook.presto.spi.StandardErrorCode.AMBIGUOUS_FUNCTION_IMPLEMENTATION;
 import static com.facebook.presto.spi.StandardErrorCode.FUNCTION_IMPLEMENTATION_ERROR;
 import static com.facebook.presto.spi.StandardErrorCode.FUNCTION_IMPLEMENTATION_MISSING;
-import static com.facebook.presto.spi.type.TypeSignature.parseTypeSignature;
 import static com.facebook.presto.util.Failures.checkCondition;
-import static com.facebook.presto.util.ImmutableCollectors.toImmutableList;
-import static com.facebook.presto.util.ImmutableCollectors.toImmutableSet;
 import static com.google.common.base.Preconditions.checkArgument;
 import static java.lang.String.format;
-import static java.lang.invoke.MethodHandles.lookup;
-import static java.lang.invoke.MethodHandles.permuteArguments;
-import static java.lang.reflect.Modifier.isStatic;
-import static java.util.Arrays.asList;
 import static java.util.Objects.requireNonNull;
 
 public class ReflectionParametricScalar
         extends SqlScalarFunction
 {
-    private static final Set<OperatorType> COMPARABLE_TYPE_OPERATORS = ImmutableSet.of(EQUAL, NOT_EQUAL, HASH_CODE);
-    private static final Set<OperatorType> ORDERABLE_TYPE_OPERATORS = ImmutableSet.of(LESS_THAN, LESS_THAN_OR_EQUAL, GREATER_THAN, GREATER_THAN_OR_EQUAL, BETWEEN);
-
     // These three collections implement something similar to partial template specialization from C++, and allow more optimized implementations to be provided for specific types
 
     // These are implementations for concrete types (they have no unbound type parameters), and have the highest priority when picking an implementation
-    private final Map<Signature, Implementation> exactImplementations;
+    private final Map<Signature, ScalarImplementation> exactImplementations;
     // These are implementations for when a type parameter binds to a specific native container type
-    private final List<Implementation> specializedImplementations;
+    private final List<ScalarImplementation> specializedImplementations;
     // These are generic implementations
-    private final List<Implementation> implementations;
+    private final List<ScalarImplementation> implementations;
 
     public static class ScalarHeaderAndMethods
     {
@@ -132,6 +91,30 @@ public class ReflectionParametricScalar
             }
             return builder.build();
         }
+
+        public ScalarImplementationHeader getHeader()
+        {
+            return header;
+        }
+
+        public List<Method> getMethods()
+        {
+            return methods;
+        }
+
+        private static List<Method> findPublicMethodsWithAnnotation(Class<?> clazz, Class<?> annotationClass)
+        {
+            ImmutableList.Builder<Method> methods = ImmutableList.builder();
+            for (Method method : clazz.getMethods()) {
+                for (Annotation annotation : method.getAnnotations()) {
+                    if (annotationClass.isInstance(annotation)) {
+                        checkArgument(Modifier.isPublic(method.getModifiers()), "%s annotated with %s must be public", method.getName(), annotationClass.getSimpleName());
+                        methods.add(method);
+                    }
+                }
+            }
+            return methods.build();
+        }
     }
 
     ScalarHeader details;
@@ -139,9 +122,9 @@ public class ReflectionParametricScalar
     private ReflectionParametricScalar(
             Signature signature,
             ScalarHeader details,
-            Map<Signature, Implementation> exactImplementations,
-            List<Implementation> specializedImplementations,
-            List<Implementation> implementations)
+            Map<Signature, ScalarImplementation> exactImplementations,
+            List<ScalarImplementation> specializedImplementations,
+            List<ScalarImplementation> implementations)
     {
         super(signature);
         this.details = details;
@@ -173,15 +156,15 @@ public class ReflectionParametricScalar
     {
         Signature boundSignature = SignatureBinder.bindVariables(getSignature(), boundVariables, arity);
         if (exactImplementations.containsKey(boundSignature)) {
-            Implementation implementation = exactImplementations.get(boundSignature);
-            MethodHandleAndConstructor methodHandleAndConstructor = implementation.specialize(boundSignature, boundVariables, typeManager, functionRegistry);
+            ScalarImplementation implementation = exactImplementations.get(boundSignature);
+            ScalarImplementation.MethodHandleAndConstructor methodHandleAndConstructor = implementation.specialize(boundSignature, boundVariables, typeManager, functionRegistry);
             checkCondition(methodHandleAndConstructor != null, FUNCTION_IMPLEMENTATION_ERROR, String.format("Exact implementation of %s do not match expected java types.", boundSignature.getName()));
             return new ScalarFunctionImplementation(implementation.isNullable(), implementation.getNullableArguments(), methodHandleAndConstructor.getMethodHandle(), methodHandleAndConstructor.getConstructor(), isDeterministic());
         }
 
         ScalarFunctionImplementation selectedImplementation = null;
-        for (Implementation implementation : specializedImplementations) {
-            MethodHandleAndConstructor methodHandle = implementation.specialize(boundSignature, boundVariables, typeManager, functionRegistry);
+        for (ScalarImplementation implementation : specializedImplementations) {
+            ScalarImplementation.MethodHandleAndConstructor methodHandle = implementation.specialize(boundSignature, boundVariables, typeManager, functionRegistry);
             if (methodHandle != null) {
                 checkCondition(selectedImplementation == null, AMBIGUOUS_FUNCTION_IMPLEMENTATION, "Ambiguous implementation for %s with bindings %s", getSignature(), boundVariables.getTypeVariables());
                 selectedImplementation = new ScalarFunctionImplementation(implementation.isNullable(), implementation.getNullableArguments(), methodHandle.getMethodHandle(), methodHandle.getConstructor(), isDeterministic());
@@ -191,8 +174,8 @@ public class ReflectionParametricScalar
             return selectedImplementation;
         }
 
-        for (Implementation implementation : implementations) {
-            MethodHandleAndConstructor methodHandle = implementation.specialize(boundSignature, boundVariables, typeManager, functionRegistry);
+        for (ScalarImplementation implementation : implementations) {
+            ScalarImplementation.MethodHandleAndConstructor methodHandle = implementation.specialize(boundSignature, boundVariables, typeManager, functionRegistry);
             if (methodHandle != null) {
                 checkCondition(selectedImplementation == null, AMBIGUOUS_FUNCTION_IMPLEMENTATION, "Ambiguous implementation for %s with bindings %s", getSignature(), boundVariables.getTypeVariables());
                 selectedImplementation = new ScalarFunctionImplementation(implementation.isNullable(), implementation.getNullableArguments(), methodHandle.getMethodHandle(), methodHandle.getConstructor(), isDeterministic());
@@ -225,15 +208,15 @@ public class ReflectionParametricScalar
 
     private static SqlScalarFunction parseScalarImplementations(ScalarHeaderAndMethods scalar, Map<Set<TypeParameter>, Constructor<?>> constructors, String objectName)
     {
-        ImmutableMap.Builder<Signature, Implementation> exactImplementations = ImmutableMap.builder();
-        ImmutableList.Builder<Implementation> specializedImplementations = ImmutableList.builder();
-        ImmutableList.Builder<Implementation> implementations = ImmutableList.builder();
+        ImmutableMap.Builder<Signature, ScalarImplementation> exactImplementations = ImmutableMap.builder();
+        ImmutableList.Builder<ScalarImplementation> specializedImplementations = ImmutableList.builder();
+        ImmutableList.Builder<ScalarImplementation> implementations = ImmutableList.builder();
         Signature signature = null;
-        ScalarImplementationHeader header = scalar.header;
+        ScalarImplementationHeader header = scalar.getHeader();
         checkArgument(!header.getName().isEmpty());
 
-        for (Method method : scalar.methods) {
-            Implementation implementation = new ImplementationParser(header.getName(), method, constructors).get();
+        for (Method method : scalar.getMethods()) {
+            ScalarImplementation implementation = ScalarImplementation.Parser.parseImplementation(header.getName(), method, constructors);
             if (implementation.getSignature().getTypeVariableConstraints().isEmpty()
                     && implementation.getSignature().getArgumentTypes().stream().noneMatch(TypeSignature::isCalculated)) {
                 exactImplementations.put(implementation.getSignature(), implementation);
@@ -253,11 +236,11 @@ public class ReflectionParametricScalar
             }
         }
 
-        Map<Signature, Implementation> exactImplementationsMap = exactImplementations.build();
+        Map<Signature, ScalarImplementation> exactImplementationsMap = exactImplementations.build();
         if (signature == null) {
             checkArgument(!exactImplementationsMap.isEmpty(), "Implementation of ScalarFunction %s must be parametric or exact implementation.", objectName);
             checkArgument(exactImplementationsMap.size() == 1, "Classes with multiple exact implementations have to have generic signature.");
-            Map.Entry<Signature, Implementation> onlyImplementation = exactImplementationsMap.entrySet().iterator().next();
+            Map.Entry<Signature, ScalarImplementation> onlyImplementation = exactImplementationsMap.entrySet().iterator().next();
             signature = onlyImplementation.getKey();
         }
 
@@ -267,20 +250,6 @@ public class ReflectionParametricScalar
                 exactImplementationsMap,
                 specializedImplementations.build(),
                 implementations.build());
-    }
-
-    private static List<Method> findPublicMethodsWithAnnotation(Class<?> clazz, Class<?> annotationClass)
-    {
-        ImmutableList.Builder<Method> methods = ImmutableList.builder();
-        for (Method method : clazz.getMethods()) {
-            for (Annotation annotation : method.getAnnotations()) {
-                if (annotationClass.isInstance(annotation)) {
-                    checkArgument(Modifier.isPublic(method.getModifiers()), "%s annotated with %s must be public", method.getName(), annotationClass.getSimpleName());
-                    methods.add(method);
-                }
-            }
-        }
-        return methods.build();
     }
 
     private static Map<Set<TypeParameter>, Constructor<?>> findConstructors(Class<?> clazz)
@@ -293,447 +262,5 @@ public class ReflectionParametricScalar
             builder.put(typeParameters, constructor);
         }
         return builder.build();
-    }
-
-    private static final class ImplementationParser
-    {
-        private final String functionName;
-        private final boolean nullable;
-        private final List<Boolean> nullableArguments = new ArrayList<>();
-        private final TypeSignature returnType;
-        private final List<TypeSignature> argumentTypes = new ArrayList<>();
-        private final List<Class<?>> argumentNativeContainerTypes = new ArrayList<>();
-        private final MethodHandle methodHandle;
-        private final List<ImplementationDependency> dependencies = new ArrayList<>();
-        private final LinkedHashSet<TypeParameter> typeParameters = new LinkedHashSet<>();
-        private final Set<String> literalParameters = new HashSet<>();
-        private final Map<String, Class<?>> specializedTypeParameters;
-        private final Optional<MethodHandle> constructorMethodHandle;
-        private final List<ImplementationDependency> constructorDependencies = new ArrayList<>();
-        private final List<LongVariableConstraint> longVariableConstraints = new ArrayList<>();
-
-        public ImplementationParser(String functionName, Method method, Map<Set<TypeParameter>, Constructor<?>> constructors)
-        {
-            this.functionName = requireNonNull(functionName, "functionName is null");
-            this.nullable = method.getAnnotation(Nullable.class) != null;
-
-            Stream.of(method.getAnnotationsByType(TypeParameter.class))
-                    .forEach(typeParameters::add);
-
-            LiteralParameters literalParametersAnnotation = method.getAnnotation(LiteralParameters.class);
-            if (literalParametersAnnotation != null) {
-                literalParameters.addAll(asList(literalParametersAnnotation.value()));
-            }
-
-            SqlType returnType = method.getAnnotation(SqlType.class);
-            requireNonNull(returnType, format("%s is missing @SqlType annotation", method));
-            this.returnType = parseTypeSignature(returnType.value(), literalParameters);
-
-            Class<?> actualReturnType = method.getReturnType();
-            if (Primitives.isWrapperType(actualReturnType)) {
-                checkArgument(nullable, "Method %s has return value with type %s that is missing @Nullable", method, actualReturnType);
-            }
-
-            Stream.of(method.getAnnotationsByType(Constraint.class))
-                    .forEach(
-                            annotation -> longVariableConstraints.add(new LongVariableConstraint(annotation.variable(), annotation.expression()))
-                    );
-
-            this.specializedTypeParameters = getDeclaredSpecializedTypeParameters(method);
-
-            parseArguments(method);
-
-            this.constructorMethodHandle = getConstructor(method, constructors);
-
-            this.methodHandle = getMethodHandle(method);
-        }
-
-        private void parseArguments(Method method)
-        {
-            ImmutableSet<String> typeParameterNames = typeParameters.stream()
-                    .map(TypeParameter::value)
-                    .collect(toImmutableSet());
-            for (int i = 0; i < method.getParameterCount(); i++) {
-                Annotation[] annotations = method.getParameterAnnotations()[i];
-                Class<?> parameterType = method.getParameterTypes()[i];
-                // Skip injected parameters
-                if (parameterType == ConnectorSession.class) {
-                    continue;
-                }
-                if (containsMetaParameter(annotations)) {
-                    checkArgument(annotations.length == 1, "Meta parameters may only have a single annotation");
-                    checkArgument(argumentTypes.isEmpty(), "Meta parameter must come before parameters");
-                    Annotation annotation = annotations[0];
-                    if (annotation instanceof TypeParameter) {
-                        checkArgument(typeParameters.contains(annotation), "Injected type parameters must be declared with @TypeParameter annotation on the method");
-                    }
-                    dependencies.add(parseDependency(annotation));
-                }
-                else {
-                    SqlType type = null;
-                    boolean nullableArgument = false;
-                    for (Annotation annotation : annotations) {
-                        if (annotation instanceof SqlType) {
-                            type = (SqlType) annotation;
-                        }
-                        if (annotation instanceof Nullable) {
-                            nullableArgument = true;
-                        }
-                    }
-                    requireNonNull(type, format("@SqlType annotation missing for argument to %s", method));
-                    if (Primitives.isWrapperType(parameterType)) {
-                        checkArgument(nullableArgument, "Method %s has parameter with type %s that is missing @Nullable", method, parameterType);
-                    }
-                    if (typeParameterNames.contains(type.value()) && !(parameterType == Object.class && nullableArgument)) {
-                        // Infer specialization on this type parameter. We don't do this for @Nullable Object because it could match a type like BIGINT
-                        Class<?> specialization = specializedTypeParameters.get(type.value());
-                        Class<?> nativeParameterType = Primitives.unwrap(parameterType);
-                        checkArgument(specialization == null || specialization.equals(nativeParameterType), "%s has conflicting specializations %s and %s", type.value(), specialization, nativeParameterType);
-                        specializedTypeParameters.put(type.value(), nativeParameterType);
-                    }
-                    argumentNativeContainerTypes.add(parameterType);
-                    argumentTypes.add(parseTypeSignature(type.value(), literalParameters));
-                    nullableArguments.add(nullableArgument);
-                }
-            }
-        }
-
-        // Find matching constructor, if this is an instance method, and populate constructorDependencies
-        private Optional<MethodHandle> getConstructor(Method method, Map<Set<TypeParameter>, Constructor<?>> constructors)
-        {
-            if (isStatic(method.getModifiers())) {
-                return Optional.empty();
-            }
-
-            Constructor<?> constructor = constructors.get(typeParameters);
-            requireNonNull(constructor, format("%s is an instance method and requires a public constructor to be declared with %s type parameters", method.getName(), typeParameters));
-            for (int i = 0; i < constructor.getParameterCount(); i++) {
-                Annotation[] annotations = constructor.getParameterAnnotations()[i];
-                checkArgument(containsMetaParameter(annotations), "Constructors may only have meta parameters");
-                checkArgument(annotations.length == 1, "Meta parameters may only have a single annotation");
-                Annotation annotation = annotations[0];
-                if (annotation instanceof TypeParameter) {
-                    checkArgument(typeParameters.contains(annotation), "Injected type parameters must be declared with @TypeParameter annotation on the constructor");
-                }
-                constructorDependencies.add(parseDependency(annotation));
-            }
-            try {
-                return Optional.of(lookup().unreflectConstructor(constructor));
-            }
-            catch (IllegalAccessException e) {
-                throw new PrestoException(FUNCTION_IMPLEMENTATION_ERROR, e);
-            }
-        }
-
-        private Map<String, Class<?>> getDeclaredSpecializedTypeParameters(Method method)
-        {
-            Map<String, Class<?>> specializedTypeParameters = new HashMap<>();
-            TypeParameterSpecialization[] typeParameterSpecializations = method.getAnnotationsByType(TypeParameterSpecialization.class);
-            ImmutableSet<String> typeParameterNames = typeParameters.stream()
-                    .map(TypeParameter::value)
-                    .collect(toImmutableSet());
-            for (TypeParameterSpecialization specialization : typeParameterSpecializations) {
-                checkArgument(typeParameterNames.contains(specialization.name()), "%s does not match any declared type parameters (%s)", specialization.name(), typeParameters);
-                Class<?> existingSpecialization = specializedTypeParameters.get(specialization.name());
-                checkArgument(existingSpecialization == null || existingSpecialization.equals(specialization.nativeContainerType()), "%s has conflicting specializations %s and %s", specialization.name(), existingSpecialization, specialization.nativeContainerType());
-                specializedTypeParameters.put(specialization.name(), specialization.nativeContainerType());
-            }
-            return specializedTypeParameters;
-        }
-
-        private MethodHandle getMethodHandle(Method method)
-        {
-            MethodHandle methodHandle;
-            try {
-                methodHandle = lookup().unreflect(method);
-            }
-            catch (IllegalAccessException e) {
-                throw new PrestoException(FUNCTION_IMPLEMENTATION_ERROR, e);
-            }
-            if (!isStatic(method.getModifiers())) {
-                // Re-arrange the parameters, so that the "this" parameter is after the meta parameters
-                int[] permutedIndices = new int[methodHandle.type().parameterCount()];
-                permutedIndices[0] = dependencies.size();
-                MethodType newType = methodHandle.type().changeParameterType(dependencies.size(), methodHandle.type().parameterType(0));
-                for (int i = 0; i < dependencies.size(); i++) {
-                    permutedIndices[i + 1] = i;
-                    newType = newType.changeParameterType(i, methodHandle.type().parameterType(i + 1));
-                }
-                for (int i = dependencies.size() + 1; i < permutedIndices.length; i++) {
-                    permutedIndices[i] = i;
-                }
-                methodHandle = permuteArguments(methodHandle, newType, permutedIndices);
-            }
-            return methodHandle;
-        }
-
-        private static List<TypeVariableConstraint> createTypeVariableConstraints(Iterable<TypeParameter> typeParameters, List<ImplementationDependency> dependencies)
-        {
-            Set<String> orderableRequired = new HashSet<>();
-            Set<String> comparableRequired = new HashSet<>();
-            for (ImplementationDependency dependency : dependencies) {
-                if (dependency instanceof OperatorImplementationDependency) {
-                    OperatorType operator = ((OperatorImplementationDependency) dependency).getOperator();
-                    if (operator == CAST) {
-                        continue;
-                    }
-                    Set<String> argumentTypes = ((OperatorImplementationDependency) dependency).getSignature().getArgumentTypes().stream()
-                            .map(TypeSignature::getBase)
-                            .collect(toImmutableSet());
-                    checkArgument(argumentTypes.size() == 1, "Operator dependency must only have arguments of a single type");
-                    String argumentType = Iterables.getOnlyElement(argumentTypes);
-                    if (COMPARABLE_TYPE_OPERATORS.contains(operator)) {
-                        comparableRequired.add(argumentType);
-                    }
-                    if (ORDERABLE_TYPE_OPERATORS.contains(operator)) {
-                        orderableRequired.add(argumentType);
-                    }
-                }
-            }
-            ImmutableList.Builder<TypeVariableConstraint> typeVariableConstraints = ImmutableList.builder();
-            for (TypeParameter typeParameter : typeParameters) {
-                String name = typeParameter.value();
-                if (orderableRequired.contains(name)) {
-                    typeVariableConstraints.add(orderableTypeParameter(name));
-                }
-                else if (comparableRequired.contains(name)) {
-                    typeVariableConstraints.add(comparableTypeParameter(name));
-                }
-                else {
-                    typeVariableConstraints.add(typeVariable(name));
-                }
-            }
-            return typeVariableConstraints.build();
-        }
-
-        private static ImplementationDependency parseDependency(Annotation annotation)
-        {
-            if (annotation instanceof TypeParameter) {
-                return new TypeImplementationDependency(((TypeParameter) annotation).value());
-            }
-            if (annotation instanceof OperatorDependency) {
-                OperatorDependency operator = (OperatorDependency) annotation;
-                return new OperatorImplementationDependency(
-                        operator.operator(),
-                        parseTypeSignature(operator.returnType()),
-                        asList(operator.argumentTypes()).stream().map(TypeSignature::parseTypeSignature).collect(toImmutableList()));
-            }
-            throw new IllegalArgumentException("Unsupported annotation " + annotation.getClass().getSimpleName());
-        }
-
-        private static boolean containsMetaParameter(Annotation[] annotations)
-        {
-            for (Annotation annotation : annotations) {
-                if (annotation instanceof TypeParameter) {
-                    return true;
-                }
-                if (annotation instanceof OperatorDependency) {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        public Implementation get()
-        {
-            Signature signature = new Signature(
-                    functionName,
-                    SCALAR,
-                    createTypeVariableConstraints(typeParameters, dependencies),
-                    longVariableConstraints,
-                    returnType,
-                    argumentTypes,
-                    false);
-            return new Implementation(
-                    signature,
-                    nullable,
-                    nullableArguments,
-                    methodHandle,
-                    dependencies,
-                    constructorMethodHandle,
-                    constructorDependencies,
-                    argumentNativeContainerTypes,
-                    specializedTypeParameters);
-        }
-    }
-
-    private static final class Implementation
-    {
-        private final Signature signature;
-        private final boolean nullable;
-        private final List<Boolean> nullableArguments;
-        private final MethodHandle methodHandle;
-        private final List<ImplementationDependency> dependencies;
-        private final Optional<MethodHandle> constructor;
-        private final List<ImplementationDependency> constructorDependencies;
-        private final List<Class<?>> argumentNativeContainerTypes;
-        private final Map<String, Class<?>> specializedTypeParameters;
-
-        private Implementation(
-                Signature signature,
-                boolean nullable,
-                List<Boolean> nullableArguments,
-                MethodHandle methodHandle,
-                List<ImplementationDependency> dependencies,
-                Optional<MethodHandle> constructor,
-                List<ImplementationDependency> constructorDependencies,
-                List<Class<?>> argumentNativeContainerTypes,
-                Map<String, Class<?>> specializedTypeParameters)
-        {
-            this.signature = requireNonNull(signature, "signature is null");
-            this.nullable = nullable;
-            this.nullableArguments = ImmutableList.copyOf(requireNonNull(nullableArguments, "nullableArguments is null"));
-            this.methodHandle = requireNonNull(methodHandle, "methodHandle is null");
-            this.dependencies = ImmutableList.copyOf(requireNonNull(dependencies, "dependencies is null"));
-            this.constructor = requireNonNull(constructor, "constructor is null");
-            this.constructorDependencies = ImmutableList.copyOf(requireNonNull(constructorDependencies, "constructorDependencies is null"));
-            this.argumentNativeContainerTypes = ImmutableList.copyOf(requireNonNull(argumentNativeContainerTypes, "argumentNativeContainerTypes is null"));
-            this.specializedTypeParameters = ImmutableMap.copyOf(requireNonNull(specializedTypeParameters, "specializedTypeParameters is null"));
-        }
-
-        public MethodHandleAndConstructor specialize(Signature boundSignature, BoundVariables boundVariables, TypeManager typeManager, FunctionRegistry functionRegistry)
-        {
-            for (Map.Entry<String, Class<?>> entry : specializedTypeParameters.entrySet()) {
-                if (!entry.getValue().isAssignableFrom(boundVariables.getTypeVariable(entry.getKey()).getJavaType())) {
-                    return null;
-                }
-            }
-            Class<?> returnContainerType = getNullAwareContainerType(typeManager.getType(boundSignature.getReturnType()).getJavaType(), nullable);
-            if (!returnContainerType.equals(methodHandle.type().returnType())) {
-                return null;
-            }
-            for (int i = 0; i < boundSignature.getArgumentTypes().size(); i++) {
-                Class<?> argumentContainerType = getNullAwareContainerType(typeManager.getType(boundSignature.getArgumentTypes().get(i)).getJavaType(), nullableArguments.get(i));
-                if (!argumentNativeContainerTypes.get(i).isAssignableFrom(argumentContainerType)) {
-                    return null;
-                }
-            }
-            MethodHandle methodHandle = this.methodHandle;
-            for (ImplementationDependency dependency : dependencies) {
-                methodHandle = methodHandle.bindTo(dependency.resolve(boundVariables.getTypeVariables(), typeManager, functionRegistry));
-            }
-            MethodHandle constructor = null;
-            if (this.constructor.isPresent()) {
-                constructor = this.constructor.get();
-                for (ImplementationDependency dependency : constructorDependencies) {
-                    constructor = constructor.bindTo(dependency.resolve(boundVariables.getTypeVariables(), typeManager, functionRegistry));
-                }
-            }
-            return new MethodHandleAndConstructor(methodHandle, Optional.ofNullable(constructor));
-        }
-
-        private static Class<?> getNullAwareContainerType(Class<?> clazz, boolean nullable)
-        {
-            if (nullable) {
-                return Primitives.wrap(clazz);
-            }
-            return clazz;
-        }
-
-        public boolean hasSpecializedTypeParameters()
-        {
-            return !specializedTypeParameters.isEmpty();
-        }
-
-        public Signature getSignature()
-        {
-            return signature;
-        }
-
-        public boolean isNullable()
-        {
-            return nullable;
-        }
-
-        public List<Boolean> getNullableArguments()
-        {
-            return nullableArguments;
-        }
-
-        public MethodHandle getMethodHandle()
-        {
-            return methodHandle;
-        }
-
-        public List<ImplementationDependency> getDependencies()
-        {
-            return dependencies;
-        }
-    }
-
-    private static final class MethodHandleAndConstructor
-    {
-        private final MethodHandle methodHandle;
-        private final Optional<MethodHandle> constructor;
-
-        public MethodHandleAndConstructor(MethodHandle methodHandle, Optional<MethodHandle> constructor)
-        {
-            this.methodHandle = requireNonNull(methodHandle, "methodHandle is null");
-            this.constructor = requireNonNull(constructor, "constructor is null");
-        }
-
-        public MethodHandle getMethodHandle()
-        {
-            return methodHandle;
-        }
-
-        public Optional<MethodHandle> getConstructor()
-        {
-            return constructor;
-        }
-    }
-
-    private interface ImplementationDependency
-    {
-        Object resolve(Map<String, Type> types, TypeManager typeManager, FunctionRegistry functionRegistry);
-    }
-
-    private static final class OperatorImplementationDependency
-            implements ImplementationDependency
-    {
-        private final OperatorType operator;
-        private final Signature signature;
-
-        private OperatorImplementationDependency(OperatorType operator, TypeSignature returnType, List<TypeSignature> argumentTypes)
-        {
-            this.operator = requireNonNull(operator, "operator is null");
-            requireNonNull(returnType, "returnType is null");
-            requireNonNull(argumentTypes, "argumentTypes is null");
-            this.signature = internalOperator(operator, returnType, argumentTypes);
-        }
-
-        public OperatorType getOperator()
-        {
-            return operator;
-        }
-
-        public Signature getSignature()
-        {
-            return signature;
-        }
-
-        @Override
-        public MethodHandle resolve(Map<String, Type> types, TypeManager typeManager, FunctionRegistry functionRegistry)
-        {
-            Signature signature = SignatureBinder.bindVariables(this.signature, new BoundVariables(types, ImmutableMap.of()), this.signature.getArgumentTypes().size());
-            return functionRegistry.getScalarFunctionImplementation(signature).getMethodHandle();
-        }
-    }
-
-    private static final class TypeImplementationDependency
-            implements ImplementationDependency
-    {
-        private final TypeSignature signature;
-
-        private TypeImplementationDependency(String signature)
-        {
-            this.signature = parseTypeSignature(requireNonNull(signature, "signature is null"));
-        }
-
-        @Override
-        public Type resolve(Map<String, Type> types, TypeManager typeManager, FunctionRegistry functionRegistry)
-        {
-            return typeManager.getType(SignatureBinder.bindVariables(signature, new BoundVariables(types, ImmutableMap.of())));
-        }
     }
 }
